producer/requirements.txt
pika
python-dotenv
